zustand in own file useEffect with boolean
