+++ app/components/MainSection.tsx
@@ -6,10 +6,10 @@ import CategoryAside from './CategoryAside';
 const MainSection = () => {
   return (
     <>
-      <main className="flex justify-center w-full">
+      <div className="flex justify-center w-full">
         <CategoryAside />
         <ArticleSection />
-      </main>
+      </div>
     </>
   );
 };
